# perf(SecureRng): Alignment & Vector optimisations