--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,6 +22,7 @@ instant = { version = "0.1", features = ["wasm-bindgen"] }
 wasm-bindgen-test = "0.3"
 
 [dependencies]
+bytemuck = { version = "1", features = ["min_const_generics"], optional = true }
 getrandom = { version = "0.2" }
 rand_core = { version = "0.6", optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
@@ -36,7 +37,7 @@ harness = false
 
 [features]
 atomic = []
-secure = []
+secure = ["dep:bytemuck"]
 rand = ["dep:rand_core"]
 serialize = ["dep:serde"]