Make ChaCha8 struct smaller with Vector backed EntropyBuffer & pass better aligned data around for extra perf gains.

Optimisations takes clone time from 237ns to 73ns, fill_bytes with 24 byte array from 75ns  to 29ns, and gen_u32 from 15ns to 10ns. Nice and speedy, without needing SIMD just yet for optimisation.